+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
-## Latest - v3.10.8 (10/03/25):
+## Latest - v3.10.9 (12/04/25):
+### Enhancements:
+- [Portico] Added withstartdate and withenddate for reporting
+- [Portico] Implemented amountIndicator element 
+
+## v3.10.8 (10/03/25):
 
 ### Enhancements: